+++ Team15091/src/main/java/org/firstinspires/ftc/team15091/RobotDriver.java
@@ -6,62 +6,18 @@ import com.qualcomm.robotcore.util.ElapsedTime;
 import com.qualcomm.robotcore.util.Range;
 
 public class RobotDriver {
-    /* Declare OpMode members. */
-    private final AztecRobot _robot;
+    private Robot _robot;
     private final ElapsedTime _runtime = new ElapsedTime();
-    private final LinearOpMode _opMode;
-
+    private LinearOpMode _opMode;
     private static final double P_TURN_COEFF = 0.05d;     // Larger is more responsive, but also less stable
     private static final double P_DRIVE_COEFF = 0.09d;     // Larger is more responsive, but also less stable
     private static final double HEADING_THRESHOLD = 0.36d;      // As tight as we can make it with an integer gyro
 
-    RobotDriver(AztecRobot robot, LinearOpMode opMode) {
+    RobotDriver(Robot robot, LinearOpMode opMode) {
         _robot = robot;
         _opMode = opMode;
 
         _robot.resetDrive();
-    }
-
-    public final void gyroSlide(double speed,
-                                double distance,
-                                double angle,
-                                double timeoutS,
-                                IObjectDetector<Boolean> objectDetector) {
-
-        // Ensure that the opmode is still active
-        if (_opMode.opModeIsActive()) {
-            // Set Target and Turn On RUN_TO_POSITION
-            _robot.setDriveTarget(distance, true);
-            _robot.setDriveMode(DcMotor.RunMode.RUN_TO_POSITION);
-
-            // start motion.
-            double counter = 0d;
-            _runtime.reset();
-            double powerToSet = Math.max(speed, 0.2d);
-            _robot.setDriveVelocity(powerToSet, powerToSet, powerToSet, powerToSet);
-            ElapsedTime driveTime = new ElapsedTime();
-
-            // keep looping while we are still active, and BOTH motors are running.
-            while (_opMode.opModeIsActive() &&
-                    _runtime.seconds() < timeoutS &&
-                    _robot.isDriveBusy()) {
-                _opMode.idle();
-
-                if (objectDetector != null && objectDetector.objectDetected()) {
-                    _robot.beep();
-                    break;
-                }
-
-                _opMode.telemetry.update();
-            }
-
-            // Stop all motion;
-            _robot.setDrivePower(0d, 0d, 0d, 0d);
-
-            // Turn off RUN_TO_POSITION
-            _robot.setDriveMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
-            _robot.setDriveMode(DcMotor.RunMode.RUN_USING_ENCODER);
-        }
     }
 
     public final boolean gyroDrive(double speed,
@@ -69,7 +25,6 @@ public class RobotDriver {
                                    double angle,
                                    double timeoutS,
                                    IObjectDetector<Boolean> objectDetector) {
-
         boolean successful = true;
         // Ensure that the opmode is still active
         if (_opMode.opModeIsActive()) {
