+++ TeamPractice/src/main/res/raw/readme.md
@@ -1,0 +1,1 @@
+Place your sound files in this folder to use them as project resources.
